CRUD for User's Journal Entries

+++ src/main/java/com/curiodesk/journalapp/service/JournalEntryService.java
@@ -23,16 +23,20 @@ public class JournalEntryService {
     }
 
     public JournalEntry saveEntry(JournalEntry journalEntry) {
-//        journalEntry.setId(null);
         return journalEntryRepository.save(journalEntry);
     }
 
-    public List<JournalEntry> getAllEntries() {
-        return journalEntryRepository.findAll();
+    private Optional<JournalEntry> findById(final ObjectId journalId) {
+        return journalEntryRepository.findById(journalId);
     }
 
-    public Optional<JournalEntry> findById(ObjectId journalId) {
-        return journalEntryRepository.findById(journalId);
+    public Optional<JournalEntry> findById(final String username,
+                                           final ObjectId journalId) {
+        User user = userService.findByUsername(username);
+        List<JournalEntry> journalEntries = user.getJournalEntries();
+        return journalEntries.stream()
+                .filter(entry -> entry.getId().equals(journalId))
+                .findFirst();
     }
 
     public JournalEntry updateEntry(ObjectId journalId, JournalEntry entry) {
@@ -55,7 +59,8 @@ public class JournalEntryService {
     }
 
     @Transactional
-    public JournalEntry saveJournalEntryOfUser(String username, JournalEntry journalEntry) {
+    public JournalEntry saveJournalEntryOfUser(final String username,
+                                               final JournalEntry journalEntry) {
         User user = userService.findByUsername(username);
         JournalEntry savedJournalEntry = journalEntryRepository.save(journalEntry);
         user.getJournalEntries().add(savedJournalEntry);
@@ -64,19 +69,32 @@ public class JournalEntryService {
     }
 
     @Transactional
-    public void deleteJournalEntryOfUser(final String username, final ObjectId journalId) {
-        User user = userService.findByUsername(username);
-        user.getJournalEntries().removeIf(entry -> entry.getId().equals(journalId));
-        userService.save(user);
-        journalEntryRepository.deleteById(journalId);
+    public boolean deleteJournalEntryOfUser(final String username, final ObjectId journalId) {
+        boolean removed = false;
+        try{
+            User user = userService.findByUsername(username);
+            removed = user.getJournalEntries().removeIf(entry -> entry.getId().equals(journalId));
+
+            if(removed) {
+                userService.save(user);
+                journalEntryRepository.deleteById(journalId);
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("An error occurred while deleting journal entry");
+        }
+        return removed;
     }
 
     public JournalEntry updateJournalEntryOfUser(final String username, ObjectId journalId, JournalEntry entry) {
-        JournalEntry existingEntry = findById(journalId).orElse(null);
-        if (existingEntry != null) {
-            existingEntry.setTitle(entry.getTitle());
-            existingEntry.setContent(entry.getContent());
-            return saveEntry(existingEntry);
+        User  user = userService.findByUsername(username);
+        boolean isUserJournal =  user.getJournalEntries().stream().anyMatch(journalEntry -> journalEntry.getId().equals(journalId));
+        if (isUserJournal) {
+            JournalEntry existingEntry = findById(journalId).orElse(null);
+            if (existingEntry != null) {
+                existingEntry.setTitle(entry.getTitle());
+                existingEntry.setContent(entry.getContent());
+                return saveEntry(existingEntry);
+            }
         }
         return null;
     }
